+++ web/src/user_group_create_members.ts
@@ -38,11 +38,6 @@ function add_members(user_ids: number[], subgroup_ids: number[]): void {
     redraw_member_list();
 }
 
-function add_all_users(): void {
-    const user_ids = user_group_create_members_data.get_all_user_ids();
-    add_members(user_ids, []);
-}
-
 function soft_remove_user_id(user_id: number): void {
     user_group_create_members_data.soft_remove_user_id(user_id);
     redraw_member_list();
@@ -91,39 +86,44 @@ function build_pill_widget({$parent_container}: {$parent_container: JQuery}): vo
 }
 
 export function create_handlers($container: JQuery): void {
-    $container.on("click", ".add_all_users_to_user_group", (e) => {
+    $container.on("click", ".remove_potential_subscriber", function (this: HTMLElement, e) {
         e.preventDefault();
-        add_all_users();
-        $(".add-user-list-filter").trigger("focus");
-    });
-
-    $container.on("click", ".remove_potential_subscriber", (e) => {
-        e.preventDefault();
-        const $elem = $(e.target);
-        const user_id = Number.parseInt($elem.attr("data-user-id")!, 10);
+        const $subscriber_row = $(this).closest(".settings-subscriber-row");
+        const user_id = Number.parseInt($subscriber_row.attr("data-user-id")!, 10);
         soft_remove_user_id(user_id);
     });
 
-    $container.on("click", ".undo_soft_removed_potential_subscriber", (e) => {
-        e.preventDefault();
-        const $elem = $(e.target);
-        const user_id = Number.parseInt($elem.attr("data-user-id")!, 10);
-        undo_soft_remove_user_id(user_id);
-    });
+    $container.on(
+        "click",
+        ".undo_soft_removed_potential_subscriber",
+        function (this: HTMLElement, e) {
+            e.preventDefault();
+            const $subscriber_row = $(this).closest(".settings-subscriber-row");
+            const user_id = Number.parseInt($subscriber_row.attr("data-user-id")!, 10);
+            undo_soft_remove_user_id(user_id);
+        },
+    );
 
-    $container.on("click", ".remove_potential_subgroup", (e) => {
+    $container.on("click", ".remove_potential_subgroup", function (this: HTMLElement, e) {
         e.preventDefault();
-        const $elem = $(e.target);
-        const subgroup_id = Number.parseInt($elem.attr("data-group-id")!, 10);
+        const $user_group_subgroup_row = $(this).closest(".user-group-subgroup-row");
+        const subgroup_id = Number.parseInt($user_group_subgroup_row.attr("data-group-id")!, 10);
         soft_remove_subgroup_id(subgroup_id);
     });
 
-    $container.on("click", ".undo_soft_removed_potential_subgroup", (e) => {
-        e.preventDefault();
-        const $elem = $(e.target);
-        const user_id = Number.parseInt($elem.attr("data-group-id")!, 10);
-        undo_soft_remove_subgroup_id(user_id);
-    });
+    $container.on(
+        "click",
+        ".undo_soft_removed_potential_subgroup",
+        function (this: HTMLElement, e) {
+            e.preventDefault();
+            const $user_group_subgroup_row = $(this).closest(".user-group-subgroup-row");
+            const subgroup_id = Number.parseInt(
+                $user_group_subgroup_row.attr("data-group-id")!,
+                10,
+            );
+            undo_soft_remove_subgroup_id(subgroup_id);
+        },
+    );
 }
 
 export function build_widgets(): void {