+++ zerver/tests/test_events.py
@@ -112,6 +112,7 @@ from zerver.actions.streams import (
     do_change_subscription_property,
     do_deactivate_stream,
     do_rename_stream,
+    do_unarchive_stream,
 )
 from zerver.actions.submessage import do_add_submessage
 from zerver.actions.typing import (
@@ -127,6 +128,7 @@ from zerver.actions.user_groups import (
     check_add_user_group,
     do_change_user_group_permission_setting,
     do_deactivate_user_group,
+    do_reactivate_user_group,
     do_update_user_group_description,
     do_update_user_group_name,
     remove_subgroups_from_user_group,
@@ -362,6 +364,7 @@ class BaseAction(ZulipTestCase):
                 linkifier_url_template=linkifier_url_template,
                 user_list_incomplete=user_list_incomplete,
                 include_deactivated_groups=include_deactivated_groups,
+                archived_channels=archived_channels,
             )
         )
 
@@ -1463,7 +1466,7 @@ class NormalActionsTest(BaseAction):
                 invite_expires_in_minutes=invite_expires_in_minutes,
             )
 
-        with self.verify_action(num_events=2) as events:
+        with self.verify_action(num_events=3) as events:
             do_deactivate_user(user_profile, acting_user=None)
         check_invites_changed("events[0]", events[0])
 
@@ -2215,14 +2218,26 @@ class NormalActionsTest(BaseAction):
             remove_subgroups_from_user_group(backend, [api_design], acting_user=None)
         check_user_group_remove_subgroups("events[0]", events[0])
 
-        # Test deactivate event
+        # Test deactivate and reactivate events
         with self.verify_action() as events:
             do_deactivate_user_group(backend, acting_user=None)
         check_user_group_remove("events[0]", events[0])
 
+        with self.verify_action() as events:
+            do_reactivate_user_group(backend, acting_user=None)
+        check_user_group_add("events[0]", events[0])
+
         with self.verify_action(include_deactivated_groups=True) as events:
             do_deactivate_user_group(api_design, acting_user=None)
         check_user_group_update("events[0]", events[0], {"deactivated"})
+        self.assertTrue(events[0]["data"]["deactivated"])
+
+        with self.verify_action(include_deactivated_groups=True) as events:
+            do_reactivate_user_group(api_design, acting_user=None)
+        check_user_group_update("events[0]", events[0], {"deactivated"})
+        self.assertFalse(events[0]["data"]["deactivated"])
+
+        do_deactivate_user_group(api_design, acting_user=None)
 
         with self.verify_action(num_events=0, state_change_expected=False):
             do_update_user_group_name(api_design, "api-deisgn-team", acting_user=None)
@@ -3471,9 +3486,10 @@ class NormalActionsTest(BaseAction):
             hamletcharacters_group, "can_mention_group", setting_group, acting_user=None
         )
 
-        with self.verify_action(num_events=1) as events:
+        with self.verify_action(num_events=2) as events:
             do_deactivate_user(user_profile, acting_user=None)
-        check_realm_user_update("events[0]", events[0], "is_active")
+        check_subscription_peer_remove("events[0]", events[0])
+        check_realm_user_update("events[1]", events[1], "is_active")
 
         do_reactivate_user(user_profile, acting_user=None)
         self.set_up_db_for_testing_user_access()
@@ -3482,9 +3498,10 @@ class NormalActionsTest(BaseAction):
         # Test that users who can access the deactivated user
         # do not receive the 'user_group/remove_members' event.
         user_profile = self.example_user("cordelia")
-        with self.verify_action(num_events=1) as events:
+        with self.verify_action(num_events=2) as events:
             do_deactivate_user(user_profile, acting_user=None)
-        check_realm_user_update("events[0]", events[0], "is_active")
+        check_subscription_peer_remove("events[0]", events[0])
+        check_realm_user_update("events[1]", events[1], "is_active")
 
         do_reactivate_user(user_profile, acting_user=None)
 
@@ -3554,15 +3571,20 @@ class NormalActionsTest(BaseAction):
         # Guest loses access to deactivated user if the user
         # was not involved in DMs.
         user_profile = self.example_user("hamlet")
-        with self.verify_action(num_events=5) as events:
+        # User is in the same channel as guest, but not in DMs.
+        self.make_stream("Test new stream")
+        self.subscribe(user_profile, "Test new stream")
+        self.subscribe(self.user_profile, "Test new stream")
+        with self.verify_action(num_events=6) as events:
             do_deactivate_user(user_profile, acting_user=None)
-        check_user_group_remove_members("events[0]", events[0])
+        check_subscription_peer_remove("events[0]", events[0])
         check_user_group_remove_members("events[1]", events[1])
         check_user_group_remove_members("events[2]", events[2])
-        check_user_group_update("events[3]", events[3], {"can_mention_group"})
-        check_realm_user_remove("events[4]]", events[4])
+        check_user_group_remove_members("events[3]", events[3])
+        check_user_group_update("events[4]", events[4], {"can_mention_group"})
+        check_realm_user_remove("events[5]]", events[5])
         self.assertEqual(
-            events[3]["data"]["can_mention_group"],
+            events[4]["data"]["can_mention_group"],
             UserGroupMembersDict(direct_members=[], direct_subgroups=[members_group.id]),
         )
 
@@ -3751,7 +3773,19 @@ class NormalActionsTest(BaseAction):
                 include_streams=include_streams, archived_channels=True
             ) as events:
                 do_deactivate_stream(stream, acting_user=None)
+            check_stream_update("events[0]", events[0])
+            self.assertEqual(events[0]["stream_id"], stream.id)
+            self.assertEqual(events[0]["property"], "is_archived")
+            self.assertEqual(events[0]["value"], True)
+
+            do_unarchive_stream(stream, stream.name, acting_user=None)
+
+            with self.verify_action(
+                include_streams=include_streams, archived_channels=False
+            ) as events:
+                do_deactivate_stream(stream, acting_user=None)
             check_stream_delete("events[0]", events[0])
+            self.assertEqual(events[0]["stream_ids"], [stream.id])
 
     def test_admin_deactivate_unsubscribed_stream(self) -> None:
         self.set_up_db_for_testing_user_access()
@@ -3768,7 +3802,37 @@ class NormalActionsTest(BaseAction):
 
         with self.verify_action(num_events=1, archived_channels=True) as events:
             do_deactivate_stream(stream, acting_user=iago)
+        check_stream_update("events[0]", events[0])
+        self.assertEqual(events[0]["stream_id"], stream.id)
+        self.assertEqual(events[0]["property"], "is_archived")
+        self.assertEqual(events[0]["value"], True)
+
+        do_unarchive_stream(stream, stream.name, acting_user=iago)
+
+        with self.verify_action(num_events=1, archived_channels=False) as events:
+            do_deactivate_stream(stream, acting_user=iago)
         check_stream_delete("events[0]", events[0])
+        self.assertEqual(events[0]["stream_ids"], [stream.id])
+
+    def test_unarchiving_stream(self) -> None:
+        iago = self.example_user("iago")
+        stream = self.make_stream("test_stream")
+        do_deactivate_stream(stream, acting_user=iago)
+
+        with self.verify_action(num_events=1, archived_channels=False) as events:
+            do_unarchive_stream(stream, stream.name, acting_user=iago)
+        check_stream_create("events[0]", events[0])
+        self.assert_length(events[0]["streams"], 1)
+        self.assertEqual(events[0]["streams"][0]["stream_id"], stream.id)
+
+        do_deactivate_stream(stream, acting_user=iago)
+
+        with self.verify_action(num_events=1, archived_channels=True) as events:
+            do_unarchive_stream(stream, stream.name, acting_user=iago)
+        check_stream_update("events[0]", events[0])
+        self.assertEqual(events[0]["stream_id"], stream.id)
+        self.assertEqual(events[0]["property"], "is_archived")
+        self.assertEqual(events[0]["value"], False)
 
     def test_user_losing_access_on_deactivating_stream(self) -> None:
         self.set_up_db_for_testing_user_access()
@@ -3784,7 +3848,7 @@ class NormalActionsTest(BaseAction):
 
         with self.verify_action(num_events=2, archived_channels=True) as events:
             do_deactivate_stream(stream, acting_user=None)
-        check_stream_delete("events[0]", events[0])
+        check_stream_update("events[0]", events[0])
 
         # Test that if the subscribers of deactivated stream are involved in
         # DMs with guest, then the guest does not get "remove" event for them.
@@ -3798,7 +3862,7 @@ class NormalActionsTest(BaseAction):
 
         with self.verify_action(num_events=2, archived_channels=True) as events:
             do_deactivate_stream(stream, acting_user=None)
-        check_stream_delete("events[0]", events[0])
+        check_stream_update("events[0]", events[0])
 
     def test_subscribe_other_user_never_subscribed(self) -> None:
         for i, include_streams in enumerate([True, False]):
@@ -4918,6 +4982,38 @@ class SubscribeActionTest(BaseAction):
                 is_web_public=False,
                 acting_user=iago,
             )
+
+        # Check updating description, stream permission for an unsubscribed streams.
+        self.user_profile = self.example_user("hamlet")
+        self.unsubscribe(self.example_user("hamlet"), stream.name)
+        with self.verify_action(include_subscribers=include_subscribers, num_events=1) as events:
+            do_change_stream_description(
+                stream, "description", acting_user=self.example_user("hamlet")
+            )
+        check_stream_update("events[0]", events[0])
+
+        with self.verify_action(include_subscribers=include_subscribers, num_events=1) as events:
+            do_change_stream_permission(
+                stream,
+                invite_only=False,
+                history_public_to_subscribers=True,
+                is_web_public=True,
+                acting_user=iago,
+            )
+        check_stream_update("events[0]", events[0])
+
+        with self.verify_action(include_subscribers=include_subscribers, num_events=1) as events:
+            do_change_stream_permission(
+                stream,
+                invite_only=True,
+                history_public_to_subscribers=False,
+                is_web_public=False,
+                acting_user=iago,
+            )
+        check_stream_update("events[0]", events[0])
+
+        # Subscribe the user again for further tests.
+        self.subscribe(self.example_user("hamlet"), stream.name)
 
         self.user_profile = self.example_user("hamlet")
         with self.verify_action(include_subscribers=include_subscribers, num_events=2) as events: